Setup for server tests (#1)
* setup jest and supertest:
- includes scripts in server, client and root package.json files

* refactor: create shared azureClient and fileUtils for upload and delete routes

* refactor: separate app and index for easier testing

* improve error handling in azureClient

* add index.ts test

* create app.ts tests

* add filesRoutes test

* add holding test files for upload and delete - to be completed after Azure connection setup

* add fileUtils tests

* refactor azureClient as a singleton class

* modify azureClient to accept credentials for testing

* add basic azureClient tests

+++ server/src/routes/deleteRoutes.ts
@@ -1,57 +1,32 @@
-import express from 'express';
-import fs from 'fs';
-import {
-  BlobServiceClient,
-  StorageSharedKeyCredential,
-} from '@azure/storage-blob';
-import { FileMetadata } from '@shared/types/files';
+import express from "express";
+import azureBlobService from "../utils/azureClient";
+import { loadFilesData, saveFilesData } from "../utils/fileUtils";
+import { FileMetadata } from "@shared/types/files";
 
 const router = express.Router();
 
-const sharedKeyCredential = new StorageSharedKeyCredential(
-  process.env.AZURE_STORAGE_ACCOUNT_NAME!,
-  process.env.AZURE_STORAGE_ACCOUNT_KEY!
-);
-
-const blobServiceClient = new BlobServiceClient(
-  `https://${process.env.AZURE_STORAGE_ACCOUNT_NAME}.blob.core.windows.net`,
-  sharedKeyCredential
-);
-
-const containerClient = blobServiceClient.getContainerClient(
-  process.env.AZURE_CONTAINER_NAME!
-);
-
-const filesDataPath = './filesData.json';
-
-const loadFilesData = (): FileMetadata[] => {
-  if (fs.existsSync(filesDataPath)) {
-    const data = fs.readFileSync(filesDataPath);
-    return JSON.parse(data.toString()) as FileMetadata[];
-  }
-  return [];
-};
-
-const saveFilesData = (files: FileMetadata[]) => {
-  fs.writeFileSync(filesDataPath, JSON.stringify(files, null, 2));
-};
-
 let files: FileMetadata[] = loadFilesData();
 
-router.delete('/:key', async (req, res) => {
+router.delete("/:key", async (req, res) => {
   const fileKey = req.params.key as string;
 
   try {
+    const containerClient = azureBlobService.getContainerClient();
+
+    if (!containerClient) {
+      throw new Error("Container client not initialized.");
+    }
+
     const blockBlobClient = containerClient.getBlockBlobClient(fileKey);
     await blockBlobClient.delete();
 
     files = files.filter((file: FileMetadata) => file.key !== fileKey);
     saveFilesData(files);
 
-    res.status(200).send('File deleted successfully.');
+    res.status(200).send("File deleted successfully.");
   } catch (err) {
-    console.error('Error deleting file:', err);
-    res.status(500).send('Failed to delete file.');
+    console.error("Error deleting file:", err);
+    res.status(500).send("Failed to delete file.");
   }
 });
