+++ server/src/routes/uploadRoutes.ts
@@ -1,47 +1,12 @@
 import express from "express";
 import multer from "multer";
 import fs from "fs";
-import dotenv from "dotenv";
-import path from "path";
-import {
-  BlobServiceClient,
-  StorageSharedKeyCredential,
-} from "@azure/storage-blob";
+import azureBlobService from "../utils/azureClient"; // Import the class instance
+import { loadFilesData, saveFilesData } from "../utils/fileUtils";
 import { FileMetadata } from "@shared/types/files";
-
-dotenv.config();
 
 const router = express.Router();
 const upload = multer({ dest: "uploads/" });
-
-const sharedKeyCredential = new StorageSharedKeyCredential(
-  process.env.AZURE_STORAGE_ACCOUNT_NAME!,
-  process.env.AZURE_STORAGE_ACCOUNT_KEY!
-);
-
-const blobServiceClient = new BlobServiceClient(
-  `https://${process.env.AZURE_STORAGE_ACCOUNT_NAME}.blob.core.windows.net`,
-  sharedKeyCredential
-);
-
-const containerClient = blobServiceClient.getContainerClient(
-  process.env.AZURE_CONTAINER_NAME!
-);
-
-const filesDataPath = path.join(__dirname, "../filesData.json");
-
-const loadFilesData = (): FileMetadata[] => {
-  if (fs.existsSync(filesDataPath)) {
-    const data = fs.readFileSync(filesDataPath);
-    return JSON.parse(data.toString()) as FileMetadata[];
-  } else {
-    return [];
-  }
-};
-
-const saveFilesData = (files: FileMetadata[]) => {
-  fs.writeFileSync(filesDataPath, JSON.stringify(files, null, 2));
-};
 
 let files: FileMetadata[] = loadFilesData();
 
@@ -53,11 +18,17 @@ router.post("/", upload.single("file"), async (req, res) => {
 
   if (req.file) {
     try {
+      const containerClient = azureBlobService.getContainerClient();
+
+      if (!containerClient) {
+        throw new Error("Container client not initialized.");
+      }
+
       const blobName = req.file.filename;
       const blockBlobClient = containerClient.getBlockBlobClient(blobName);
 
       await blockBlobClient.uploadFile(req.file.path);
-      fs.unlinkSync(req.file.path); // remove the file locally after upload
+      fs.unlinkSync(req.file.path);
 
       files.push({ name: fileName, key: blobName });
       saveFilesData(files);
